# openapi.json
{"components":{"schemas":{"NotificationView":{"description":"Vista de lectura de una notificaci\u00f3n devuelta por la API.","properties":{"_id":{"description":"Identificador \u00fanico de la notificaci\u00f3n.","title":"Id","type":"string"},"createdAt":{"description":"Fecha y hora de creaci\u00f3n de la notificaci\u00f3n.","format":"date-time","title":"Createdat","type":"string"},"email":{"anyOf":[{"format":"email","type":"string"},{"type":"null"}],"default":null,"description":"Correo electr\u00f3nico asociado al usuario.","title":"Email"},"email_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Motivo del env\u00edo o no del correo electr\u00f3nico.","title":"Email Reason"},"email_sent":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":null,"description":"Indica si se ha enviado un correo electr\u00f3nico al usuario.","title":"Email Sent"},"message":{"description":"Contenido principal de la notificaci\u00f3n.","title":"Message","type":"string"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"description":"Informaci\u00f3n adicional asociada a la notificaci\u00f3n.","title":"Metadata"},"plan":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Plan de suscripci\u00f3n del usuario.","title":"Plan"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"T\u00edtulo de la notificaci\u00f3n.","title":"Title"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Tipo de evento que origin\u00f3 la notificaci\u00f3n.","title":"Type"},"userId":{"description":"IBAN identificador del usuario propietario de la notificaci\u00f3n.","title":"Userid","type":"string"}},"required":["_id","userId","message","createdAt"],"title":"NotificationView","type":"object"}}},"info":{"title":"Notifications Service","version":"0.1.0"},"openapi":"3.1.0","paths":{"/v1/notifications/":{"get":{"description":"\nEndpoint pensado para tareas de administraci\u00f3n,\nmonitorizaci\u00f3n o debugging interno.","operationId":"get_get_all_notifications","parameters":[],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NotificationView"},"type":"array"}}},"description":"Built-in mutable sequence.\n\nIf no argument is given, the constructor creates a new empty list.\nThe argument must be an iterable if specified."}},"summary":"Devuelve todas las notificaciones del sistema.","tags":["Consulta - Admin"]}},"/v1/notifications/events":{"post":{"description":"\nRecibe eventos procedentes de otros microservicios (login, pagos,\nfraude, etc.) y genera notificaciones persistidas en base de datos.\nDependiendo del tipo de evento, puede desencadenar el env\u00edo de correos\nelectr\u00f3nicos mediante SendGrid.","operationId":"post_receive_event","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"description":"Evento recibido desde otros microservicios del sistema.","properties":{"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"description":"Informaci\u00f3n contextual adicional enviada por el microservicio emisor.","title":"Metadata"},"plan":{"anyOf":[{"enum":["basico","premium","pro"],"type":"string"},{"type":"null"}],"default":null,"description":"Plan de suscripci\u00f3n del usuario en el momento del evento.","title":"Plan"},"type":{"description":"Tipo de evento que debe ser procesado por el servicio de notificaciones.","enum":["login","transaction","scheduled-payment","history-request","fraud-detected"],"title":"Type","type":"string"},"userId":{"description":"IBAN identificador del usuario que ha generado el evento.","title":"Userid","type":"string"}},"required":["userId","type"],"title":"NotificationEvent","type":"object"}}}},"responses":{"201":{"content":{"application/json":{"schema":{"description":"Vista de lectura de una notificaci\u00f3n devuelta por la API.","properties":{"_id":{"description":"Identificador \u00fanico de la notificaci\u00f3n.","title":"Id","type":"string"},"createdAt":{"description":"Fecha y hora de creaci\u00f3n de la notificaci\u00f3n.","format":"date-time","title":"Createdat","type":"string"},"email":{"anyOf":[{"format":"email","type":"string"},{"type":"null"}],"default":null,"description":"Correo electr\u00f3nico asociado al usuario.","title":"Email"},"email_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Motivo del env\u00edo o no del correo electr\u00f3nico.","title":"Email Reason"},"email_sent":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":null,"description":"Indica si se ha enviado un correo electr\u00f3nico al usuario.","title":"Email Sent"},"message":{"description":"Contenido principal de la notificaci\u00f3n.","title":"Message","type":"string"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"description":"Informaci\u00f3n adicional asociada a la notificaci\u00f3n.","title":"Metadata"},"plan":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Plan de suscripci\u00f3n del usuario.","title":"Plan"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"T\u00edtulo de la notificaci\u00f3n.","title":"Title"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Tipo de evento que origin\u00f3 la notificaci\u00f3n.","title":"Type"},"userId":{"description":"IBAN identificador del usuario propietario de la notificaci\u00f3n.","title":"Userid","type":"string"}},"required":["_id","userId","message","createdAt"],"title":"NotificationView","type":"object"}}},"description":"Vista de lectura de una notificaci\u00f3n devuelta por la API."}},"summary":"Endpoint principal del microservicio de notificaciones.","tags":["Eventos"]}},"/v1/notifications/health":{"get":{"description":"\nPensado para ser utilizado como endpoint de liveness/readiness\nen entornos Docker o Kubernetes.","operationId":"get_health_check","parameters":[],"responses":{},"summary":"Healthcheck del microservicio.","tags":["Health"]}},"/v1/notifications/test-email":{"post":{"description":"\nEnv\u00eda un correo electr\u00f3nico de prueba utilizando la configuraci\u00f3n\nactual del servicio de notificaciones.","operationId":"post_test_email","parameters":[],"responses":{},"summary":"Endpoint de prueba para verificar la integraci\u00f3n con SendGrid.","tags":["Debug"]}},"/v1/notifications/user/{userId}":{"get":{"description":"\nLas notificaciones se devuelven ordenadas por fecha de creaci\u00f3n,\nde la m\u00e1s reciente a la m\u00e1s antigua.","operationId":"get_get_notifications_by_user","parameters":[{"in":"path","name":"userId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NotificationView"},"type":"array"}}},"description":"Built-in mutable sequence.\n\nIf no argument is given, the constructor creates a new empty list.\nThe argument must be an iterable if specified."}},"summary":"Devuelve todas las notificaciones asociadas a un usuario concreto.","tags":["Consulta - Usuario"]}},"/v1/notifications/{notification_id}":{"delete":{"description":"\nSi la notificaci\u00f3n no existe, se devuelve un error 404.","operationId":"delete_delete_notification","parameters":[{"in":"path","name":"notification_id","required":true,"schema":{"type":"string"}}],"responses":{},"summary":"Elimina una notificaci\u00f3n existente por su identificador.","tags":["Admin - Delete"]},"put":{"description":"\nSolo se modificar\u00e1n los campos presentes en el payload.","operationId":"put_update_notification","parameters":[{"in":"path","name":"notification_id","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"description":"Payload utilizado para actualizar parcialmente una notificaci\u00f3n existente.","properties":{"email_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Motivo actualizado del env\u00edo o no del correo electr\u00f3nico.","title":"Email Reason"},"email_sent":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":null,"description":"Indica si el correo electr\u00f3nico ha sido enviado.","title":"Email Sent"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Nuevo contenido del mensaje.","title":"Message"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"description":"Informaci\u00f3n contextual adicional actualizada.","title":"Metadata"},"plan":{"anyOf":[{"enum":["basico","premium","pro"],"type":"string"},{"type":"null"}],"default":null,"description":"Nuevo plan de suscripci\u00f3n del usuario.","title":"Plan"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Nuevo t\u00edtulo de la notificaci\u00f3n.","title":"Title"},"type":{"anyOf":[{"enum":["login","transaction","scheduled-payment","history-request","fraud-detected"],"type":"string"},{"type":"null"}],"default":null,"description":"Nuevo tipo de evento asociado a la notificaci\u00f3n.","title":"Type"}},"title":"NotificationUpdate","type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"description":"Vista de lectura de una notificaci\u00f3n devuelta por la API.","properties":{"_id":{"description":"Identificador \u00fanico de la notificaci\u00f3n.","title":"Id","type":"string"},"createdAt":{"description":"Fecha y hora de creaci\u00f3n de la notificaci\u00f3n.","format":"date-time","title":"Createdat","type":"string"},"email":{"anyOf":[{"format":"email","type":"string"},{"type":"null"}],"default":null,"description":"Correo electr\u00f3nico asociado al usuario.","title":"Email"},"email_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Motivo del env\u00edo o no del correo electr\u00f3nico.","title":"Email Reason"},"email_sent":{"anyOf":[{"type":"boolean"},{"type":"null"}],"default":null,"description":"Indica si se ha enviado un correo electr\u00f3nico al usuario.","title":"Email Sent"},"message":{"description":"Contenido principal de la notificaci\u00f3n.","title":"Message","type":"string"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"description":"Informaci\u00f3n adicional asociada a la notificaci\u00f3n.","title":"Metadata"},"plan":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Plan de suscripci\u00f3n del usuario.","title":"Plan"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"T\u00edtulo de la notificaci\u00f3n.","title":"Title"},"type":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Tipo de evento que origin\u00f3 la notificaci\u00f3n.","title":"Type"},"userId":{"description":"IBAN identificador del usuario propietario de la notificaci\u00f3n.","title":"Userid","type":"string"}},"required":["_id","userId","message","createdAt"],"title":"NotificationView","type":"object"}}},"description":"Vista de lectura de una notificaci\u00f3n devuelta por la API."}},"summary":"Actualiza los campos de una notificaci\u00f3n existente.","tags":["Notifications - Update"]}}},"tags":[{"description":"API version 1","name":"v1"}]}